vexpress: init: Setup cpufreq for TC2
As TC2 is the only vexpress CoreTile so far which can support cpufreq
it's safe to assume the sysfs entries for this don't exists on anything
other than TC2, so poking them is a harmless error on these other
CoreTiles.

We need to revisit this when other vexpress hardware supports cpufreq
but for now, lets just get TC2 working nicely.

Change-Id: I4c8bea0af4d76fd236b482c3fb1422e498fcad56
Signed-off-by: Jon Medhurst <tixy@linaro.org>