vexpress: Fix sdcard partition mounting on device-tree kernels
With device-tree kernels the sysfs path for SD cards varies
depending on device-tree layout and the actual card used. It is
therefore impossible to specify the sysfs path for the device
in vold.fstab.

However, there exists a symlink to the device which is created as
/sys/block/mmcblk0/device, so we can use that instead.

This depends on http://review.android.git.linaro.org/#change,2715
and fixes Launchpad bug 1042653.

Change-Id: I3927fe76d1af2be643aed1f6195ef204489b3ca2
Signed-off-by: Jon Medhurst <tixy@linaro.org>