VExpress: add audio support for Jelly Bean builds.
Based on the Android audio code for Samsung Tuna device.

VExpress audio hardware supports only 48000 sample rates per seconds,
different from the default Android 44100 value. To get audio to work
we need to create a resampler that does the conversion.

On Elba hardware that had a broken PCM implementation we also need
an echo reference in order to allow the audio stack to detect the
delays introduced by the system.

Change-Id: I639e21200ed0d2c79078d4cff40d9c413789f489
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>