vexpress: Setup IRQ affinity to A7s for TC2
On TC2, to maximise power savings, it helps to affine the IRQs to the
A7s. Now that the cpu to parts mapping is available via /proc/cpuinfo,
this patch introduces a script to parse this file and build a mask of
particular CPU type and change the IRQ affinity to this mask.

The patch also makes the necessary changes to platform specific init
script to call above script and affine the IRQs to the A7s.

Lastly, it changes the makefile to include the new script in the target.

Change-Id: I2fd2840acd4602d9145c7c2f2e839e5dda09ff99
Signed-off-by: Punit Agrawal <punit.agrawal@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>