vexpress: Build kernel and initrd images without U-Boot headers
On vexpress, uImage format is broken and a hindrance. It's been broken
since multiplatform support was introduced a mainline kernel release or
two ago, after which the load address and entrypoint is never specified
in the uImage header (it defaults to -EFAULT). Even if it did get set,
then it still wouldn't work for us, because A9 and TC2 require
different addresses. This problem has been worked around by hacking our
U-Boots to ignore the load address and use a hard coded value.

uImage format is also a hindrance for the other bootloaders people use.
For ARM's Boot Monitor, the user has to manually strip the U-Boot header
off files before flashing them; and for UEFI a hack was added detect the
header an ignore this.

Considering that U-Boot can also cope with loading naked zImage and
initrd files we should create these and all three bootloaders can use
them without hacks being required.

To enable a transition period for updating automated systems like LAVA
this patch includes make rules for creating uImage and uInitrd files in
addition to the new naked zImage and initrd files.

Change-Id: I35dac1fcd2ba81f802de7667c4864f03ad10304f
Signed-off-by: Jon Medhurst <tixy@linaro.org>