vexpress: Add support for using filesystems on USB
Add a new hardware variant "arm-versatileexpress-usb" which
will mount filesystems from a USB drive rather than MMC.

To select this variant, add the following to the kernel commanline:
  androidboot.hardware=arm-versatileexpress-usb

Note, when booting the USB variant, the 'sdcard' partition (partition 6)
on the MMC card will still be used as the external storage, if it is
present. It it is not present then the partition on USB will be used.

Change-Id: Iff0cbcabd0cc39d6f0cae5fc2ee604cb3c98ef5b
Signed-off-by: Jon Medhurst <tixy@linaro.org>