vexpress: Ensure interactive governor set for all CPUs
The current method of setting the interactive governor for both CPU
clusters on TC2 is not robust as it assumes logical CPU ordering makes
CPU0 on one cluster and CPU2 on another. This is normally the case but
if device-trees are modified (e.g. to accommodate IKS expectations of
CPU order) then the assumptions fall down.

Fix this by explicitly specifying the governor for all 5 CPUs.

Change-Id: I482d8b56e07fd347dd510c1109d829ed3f0db607
Signed-off-by: Jon Medhurst <tixy@linaro.org>