vexpress: Fix sdcard entries in fstab for 4.3
vold expects a path under /sys not one in /dev, and we have to give
it the partition number (6) to stop it picking the wrong one, e.g.
the first partition which is the boot partition.

Note, we use paths which are symlinks to the real /sys devices because
the real device path varies depending on which kind of SD card is used
and which vexpress CoreTile or model is being run. Also, the AOSP version
of vold doesn't handle symlinks, so this requires vold to be patched as
well. See change I68e693772e70b758dab22fcd5f0eb4b9e84a2443

Change-Id: I389a54005f38690a14c3cba16bd3fd458777ab4a
Signed-off-by: Jon Medhurst <tixy@linaro.org>