bootloader64.mk: do not remove dtc script
Removing dtc script from boot-wrapper directory is causing race
condition since both of Juice bootloader targets try to use it
when we do parallel builds.

One of the bootloader targets (fvp dtb build) will be moved to uefi
directory and will be build as a part of uefi build procedure once
we move to uefi for fvp base models.

Change-Id: I08d41818025a0b337faf782d664b8cc6bdc5732f
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>