vexpress: Mark sdcard as removable="false"
In Android 4.4 all access to external media is routed through a fuse
daemon, however, MediaProvider is still trying to identify removable
media with the FAT specific fucntion: FileUtils.getFatVolumeId.
As the fuse daemon doesn't emulate this feature access to external
media is always failing. For more, see [1].

To work around this issue we will mark the sdcard as removable="false";
which is in fact a more accurate descriptions of the state of affairs
anyway, because the backing for the 'sdcard' is a partition on the
Linaro Android image which also contains partitions for 'system', and
therefore could never safely be physically removed.

[1] http://lists.linaro.org/pipermail/linaro-android/2013-November/002335.html

Change-Id: I892fb26329eacb8ec1b0c2f7a53b468d76413115
Signed-off-by: Jon Medhurst <tixy@linaro.org>